Refactor DeepLinksService to a component
